moved arrow stepper to its own file
